Move frob norm and cholesky solve to math
